Truncate the percent_complete for consistency with display
For non-integer percentages, we'd end up with something like 8.1% but
then truncate when displaying. This would cause things like a prompt
that says a number must be at least 8 but then 8 would not be allowed
because it's less than 8.1.

+++ rollingpin/frontends.py
@@ -133,7 +133,7 @@ class HeadlessFrontend(object):
         return sum(1 for v in self.host_results.itervalues() if v)
 
     def percent_complete(self):
-        return (self.count_completed_hosts() / self.count_hosts()) * 100
+        return int((self.count_completed_hosts() / self.count_hosts()) * 100)
 
     def on_host_end(self, host, results):
         if host in self.host_results:
